refactor:move ati tests per comment review
